Moving traits into 'With' namespace.
