+++ src/Controllers/ModelMorphController.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace EricDowell\ResourceController\Controllers;
 
-use EricDowell\ResourceController\Traits\WithMorphModel;
+use EricDowell\ResourceController\Traits\With\MorphModel;
 
 abstract class ModelMorphController extends Controller
 {
-    use WithMorphModel;
+    use MorphModel;
 }
