+++ src/Controllers/ResourceModelController.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace EricDowell\ResourceController\Controllers;
 
-use EricDowell\ResourceController\Traits\WithModelResource;
+use EricDowell\ResourceController\Traits\With\ModelResource;
 
 abstract class ResourceModelController extends Controller
 {
-    use WithModelResource;
+    use ModelResource;
 }
